requirements.txt
Django==1.8.5
djangorestframework==3.2.4
djangorestframework-jwt==1.7.2
PyJWT==1.4.0
wheel==0.24.0
